+++ changelog.md
@@ -29,6 +29,7 @@
 
 ## Next
 
+* Run jshint on tests
 * Disable Travis dependency cache
 * Travis runs on new container infrastructure
 * README code examples tagged as Javascript